Module pest_meta::parser

Types and helpers for the pest’s own grammar parser.

Structs

ParserNode

The pest grammar node

ParserRule

The pest grammar rule

PestParser

Enums

ParserExpr

All possible parser expressions

Rule

Pest meta-grammar

Functions

consume_rules

Converts a parser’s result (Pairs) to an AST

parse

A helper that will parse using the pest grammar

rename_meta_rule

A helper function to rename verbose rules for the sake of better error messages