Module pest_meta::validator

Helpers for validating pest grammars that could help with debugging and provide a more user-friendly error message.

Statics

BUILTINS ğŸ”’

PEST_KEYWORDS ğŸ”’

RUST_KEYWORDS ğŸ”’

Functions

is_non_failing ğŸ”’

Checks if expr is non-failing, that is it matches any input.

is_non_progressing ğŸ”’

Checks if expr is non-progressing, that is the expression does not consume any input or any stack. This includes expressions matching the empty input, SOI and Ì€ EOI, predicates and repetitions.

left_recursion ğŸ”’

to_hash_map ğŸ”’

validate_already_defined

Validates that the given definitions do not contain any duplicate rules.

validate_ast

Validates the abstract syntax tree for common mistakes:

validate_choices ğŸ”’

validate_left_recursion ğŸ”’

validate_pairs

It checks the parsed grammar for common mistakes:

validate_pest_keywords

Validates that the given definitions do not contain any Pest keywords.

validate_repetition ğŸ”’

validate_rust_keywordsDeprecated

Validates that the given definitions do not contain any Rust keywords.

validate_undefined

Validates that the given definitions do not contain any undefined rules.

validate_whitespace_comment ğŸ”’