Crate pest_meta

pest meta

This crate parses, validates, optimizes, and converts pest’s own grammars to ASTs.

Modules

ast

Types for the pest’s abstract syntax tree.

optimizer

Different optimizations for pest’s ASTs.

parser

Types and helpers for the pest’s own grammar parser.

validator

Helpers for validating pest grammars that could help with debugging and provide a more user-friendly error message.

Functions

parse_and_optimize

Parses, validates, processes and optimizes the provided grammar.

unwrap_or_report

A helper that will unwrap the result or panic with the nicely formatted error message.