pest_meta

Module validator

Helpers for validating pest grammars that could help with debugging and provide a more user-friendly error message.

Functions

validate_already_defined

Validates that the given definitions do not contain any duplicate rules.

validate_ast

Validates the abstract syntax tree for common mistakes:

validate_pairs

It checks the parsed grammar for common mistakes:

validate_pest_keywords

Validates that the given definitions do not contain any Pest keywords.

validate_rust_keywordsDeprecated

Validates that the given definitions do not contain any Rust keywords.

validate_undefined

Validates that the given definitions do not contain any undefined rules.